move file (following namespace change)
